Move core bootstrap code out of the hook injector and into the main Agent Refactor hook injector to accept a HookController instance as a dependency Rename Injector to ClassFileHookInjector to better reflect its purpose
